requirements.txt
asgiref==3.7.2
crispy-bootstrap5==2023.10
Django==4.2.8
django-crispy-forms==2.1
mysqlclient
setuptools==68.0.0
sqlparse
wheel==0.41.2
